+++ openapi/openapi/src/main/scala/endpoints/ujson/JsonSchemas.scala
@@ -16,7 +16,7 @@ import scala.collection.mutable
 /**
   * @group interpreters
   */
-trait JsonSchemas extends algebra.JsonSchemas with TuplesSchemas {
+trait JsonSchemas extends algebra.NoDocsJsonSchemas with TuplesSchemas {
 
   trait JsonSchema[A] {
 
@@ -121,12 +121,6 @@ trait JsonSchemas extends algebra.JsonSchemas with TuplesSchemas {
       }
       val encoder = tpe.encoder
     }
-
-  def namedRecord[A](schema: Record[A], name: String): Record[A] = schema
-
-  def namedTagged[A](schema: Tagged[A], name: String): Tagged[A] = schema
-
-  def namedEnum[A](schema: Enum[A], name: String): Enum[A] = schema
 
   def lazyRecord[A](schema: => Record[A], name: String): JsonSchema[A] =
     new JsonSchema[A] {
@@ -237,66 +231,6 @@ trait JsonSchemas extends algebra.JsonSchemas with TuplesSchemas {
         }
       }
     }
-
-  def withExampleRecord[A](
-      schema: Record[A],
-      example: A
-  ): Record[A] = schema
-
-  def withExampleTagged[A](
-      schema: Tagged[A],
-      example: A
-  ): Tagged[A] = schema
-
-  def withExampleEnum[A](
-      schema: Enum[A],
-      example: A
-  ): Enum[A] = schema
-
-  def withExampleJsonSchema[A](
-      schema: JsonSchema[A],
-      example: A
-  ): JsonSchema[A] = schema
-
-  def withDescriptionRecord[A](
-      schema: Record[A],
-      description: String
-  ): Record[A] = schema
-
-  def withDescriptionTagged[A](
-      schema: Tagged[A],
-      description: String
-  ): Tagged[A] = schema
-
-  def withDescriptionEnum[A](
-      schema: Enum[A],
-      description: String
-  ): Enum[A] = schema
-
-  def withDescriptionJsonSchema[A](
-      schema: JsonSchema[A],
-      description: String
-  ): JsonSchema[A] = schema
-
-  def withTitleRecord[A](
-      schema: Record[A],
-      title: String
-  ): Record[A] = schema
-
-  def withTitleTagged[A](
-      schema: Tagged[A],
-      title: String
-  ): Tagged[A] = schema
-
-  def withTitleEnum[A](
-      schema: Enum[A],
-      title: String
-  ): Enum[A] = schema
-
-  def withTitleJsonSchema[A](
-      schema: JsonSchema[A],
-      title: String
-  ): JsonSchema[A] = schema
 
   def orFallbackToJsonSchema[A, B](
       schemaA: JsonSchema[A],
